Refactor InfinispanConfiguration into wildfly-clustering-ee-infinispan. Add convenience methods for various cache flag sets.
