+++ clustering/ee/cache/src/main/java/org/wildfly/clustering/ee/cache/CacheConfiguration.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2019, Red Hat, Inc., and individual contributors
+ * Copyright 2021, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -20,20 +20,21 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.infinispan.session;
+package org.wildfly.clustering.ee.cache;
 
-import org.infinispan.Cache;
-import org.wildfly.clustering.ee.cache.CacheProperties;
-import org.wildfly.clustering.ee.infinispan.InfinispanCacheProperties;
+import java.util.Map;
+
+import org.wildfly.clustering.ee.Batcher;
+import org.wildfly.clustering.ee.cache.tx.TransactionBatch;
 
 /**
  * @author Paul Ferraro
  */
-public interface InfinispanConfiguration {
+public interface CacheConfiguration {
 
-    <K, V> Cache<K, V> getCache();
+    <K, V> Map<K, V> getCache();
 
-    default CacheProperties getCacheProperties() {
-        return new InfinispanCacheProperties(this.getCache().getCacheConfiguration());
-    }
+    CacheProperties getCacheProperties();
+
+    Batcher<TransactionBatch> getBatcher();
 }
